feat: Upload file for external medial player
