+++ src/assets/locales/en/translation.json
@@ -235,6 +235,9 @@
       "external-media-player-play": "Play",
       "external-media-player-close-confirm": "Are you sure to stop playback?",
       "external-media-player-close-msg": "You are about to close playback",
+      "external-media-player-direct-link": "Direct link",
+      "external-media-player-upload-file": "Upload",
+      "external-media-player-upload-supported-files": "Supported files: {{files}}",
       "upload-background-image": "Upload background image"
     }
   },
